thmswtch + slpquiet -further linter fixes thmswtch -movement of menu to config file and creation of app.js for switching only
